refactor(threads): 按模块拆分 reducer 文本合并逻辑
背景：src/features/threads/hooks/useThreadsReducer.ts 超过大文件阈值（3000 行），影响可维护性与后续演进。

本次改动：

- 新增 threadReducerTextMerge.ts，承载 assistant/reasoning 文本合并与去重、snapshot/incremental 判定、边界处理等纯函数。

- useThreadsReducer.ts 保留状态与 action 分发逻辑，通过 import 调用拆分后的文本合并函数。

- 不改变既有业务行为，不调整 action 协议，不改外部调用路径。

兼容性与验证：

- pnpm tsc --noEmit 通过。

- pnpm eslint src/features/threads/hooks/useThreadsReducer.ts src/features/threads/hooks/threadReducerTextMerge.ts 通过。

- pnpm vitest run src/features/threads/hooks/useThreadsReducer.test.ts src/features/threads/hooks/useThreadItemEvents.test.ts src/features/app/hooks/useAppServerEvents.test.tsx src/features/threads/loaders/historyLoaders.test.ts 通过（161 tests）。

- node scripts/check-large-files.mjs --threshold 3000 --mode fail 通过（found=0）。
